refactor(core): centralize enums following DDD pattern
BREAKING CHANGE: Enums moved to dedicated directories for discoverability

- Create src/core/enums/ directory
  - error_code.py: ErrorCode enum (domain error codes)
  - environment.py: Environment enum (dev/test/prod/ci)
  - __init__.py: Export all core enums

- Create src/infrastructure/enums/ directory
  - infrastructure_error_code.py: InfrastructureErrorCode enum
  - __init__.py: Export all infrastructure enums

- Update imports across codebase
  - src/core/errors.py: Import ErrorCode from enums
  - src/core/config.py: Import Environment from enums
  - src/infrastructure/errors.py: Import InfrastructureErrorCode from enums
  - src/core/__init__.py: Import ErrorCode from enums

Architecture:
- Enums are Value Objects in DDD - get their own directory
- Follows industry standard (Django, FastAPI complex projects)
- Avoids circular imports (enums are leaf nodes)
- Single Responsibility Principle (one enum per file)
- Easy to discover and maintain as project scales

Testing:
- All 166 tests passing
- Type checking passes (mypy)
- Code quality checks pass

Related: Directory structure will be documented in follow-up commit
