+++ src/core/__init__.py
@@ -8,7 +8,8 @@ This module provides foundational utilities used across all architectural layers
 The core module has NO dependencies on other application layers.
 """
 
-from src.core.errors import DomainError, ErrorCode, ValidationError
+from src.core.errors import DomainError, ValidationError
+from src.core.enums import ErrorCode
 from src.core.result import Failure, Result, Success
 
 __all__ = [
