+++ src/infrastructure/enums/__init__.py
@@ -1,0 +1,13 @@
+"""Infrastructure enums package.
+
+Exports all infrastructure-level enums for convenient importing.
+
+Usage:
+    from src.infrastructure.enums import InfrastructureErrorCode
+"""
+
+from src.infrastructure.enums.infrastructure_error_code import (
+    InfrastructureErrorCode,
+)
+
+__all__ = ["InfrastructureErrorCode"]
